fixing warnings

+++ MQTTSNPacket/test/test1.c
@@ -952,7 +952,7 @@ int main(int argc, char** argv)
 
  	if (options.test_no == 0)
 	{ /* run all the tests */
- 	   	for (options.test_no = 1; options.test_no < ARRAY_SIZE(tests); ++options.test_no)
+ 	   	for (options.test_no = 1; options.test_no < (int)ARRAY_SIZE(tests); ++options.test_no)
 			rc += tests[options.test_no](options); /* return number of failures.  0 = test succeeded */
 	}
 	else
